src/requirements.txt
flask
numpy
joblib
scikit-learn
google-generativeai